third_party/WebKit/LayoutTests/external/wpt/credential-management/passwordcredential-framed-get.sub.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
function create_iframe_test(origin, expectation) {
    return function (t) {
        window.addEventListener("load", _ => {
            var iframe = document.createElement("iframe");
            iframe.src = origin + "/credential-management/support/passwordcredential-get.html";
            window.addEventListener("message", t.step_func(e => {
                if (e.source == iframe.contentWindow) {
                    assert_equals(e.data.status, expectation);
                    if (expectation == "resolved") {
                      assert_equals(e.data.exception, "NotAllowedError");
                    } else {
                      assert_equals(e.data.exception, null);
                    }
                    t.done();
                }
            }));
            document.body.appendChild(iframe);
        });
    };
}

async_test(create_iframe_test(window.origin, "resolved"), "Same-origin IFrame does not throw.");
async_test(create_iframe_test("https://{{domains[élève]}}:{{ports[https][0]}}", "rejected"), "Cross-origin IFrame throws.");
</script>
