third_party/WebKit/LayoutTests/external/wpt/credential-management/passwordcredential-framed-get.sub.https-expected.txt

This is a testharness.js-based test.
FAIL Same-origin IFrame does not throw. assert_equals: expected "resolved" but got "rejected"
PASS Cross-origin IFrame throws 'NotAllowedError'.
FAIL Same-origin IFrame in same-origin IFrame does not throw. assert_equals: expected "resolved" but got "rejected"
PASS Same-origin IFrame in same-origin IFrame throws.
PASS Cross-origin IFrame in same-origin IFrame throws.
PASS Cross-origin IFrame in same-cross-origin throws.
Harness: the test ran to completion.
